post_vault_data

Vault: Posts data to the path you specify in Vault.

post_vault_data(url = NULL, path = NULL, token = NULL, secrets = NULL)

Arguments

url
URL of the Hashicorp Vault instance.

path
path to the secret in the vault instance.

token
token for the vault instance.

secrets
secrets that are being written to vault.

Value

Allows a user to write data to Vault.

Details

This function writes secrets or data to the vault instance and gives a message if they were written correctly. You will need the Hashicorp Vault url you are using including the full path to the secret e.g. (https://url.vault.com/v1/secret/things) and lastly you will need a token that allows you to interact with Vault via the API. You can pass a list of secrets at once use 'list(one="two",three="four) and so on and so forth. This will allow you to send several secrets to Vault at once. Remember to not overwrite secrets you will want to write new secrets with the old secrets as placeholders.

Examples

if (FALSE)   post_vault_data(url,path,token,token,secrets=list(one="two",three="four"))